requirements.txt
spacy
spacy-llm
pyyaml
python-dotenv
jinja2
openai
